CDU 1ª parte

+++ src/main/java/pt/ipleiria/estg/dei/ei/dae/clinics/entities/BiometricDataIssue.java
@@ -24,10 +24,10 @@ public class BiometricDataIssue implements Serializable {
     private String name;
 
     @NotNull
-    private int min;
+    private double min;
 
     @NotNull
-    private int max;
+    private double max;
 
     @NotNull
     @ManyToOne
@@ -41,7 +41,7 @@ public class BiometricDataIssue implements Serializable {
                     "ID"))
     private List<Prescription> prescriptions;
 
-    public BiometricDataIssue(String name, int min, int max, BiometricDataType biometric_data_type) {
+    public BiometricDataIssue(String name, double min, double max, BiometricDataType biometric_data_type) {
         this.name = name;
         this.min = min;
         this.max = max;
@@ -85,19 +85,19 @@ public class BiometricDataIssue implements Serializable {
         this.name = name;
     }
 
-    public int getMin() {
+    public double getMin() {
         return min;
     }
 
-    public void setMin(int min) {
+    public void setMin(double min) {
         this.min = min;
     }
 
-    public int getMax() {
+    public double getMax() {
         return max;
     }
 
-    public void setMax(int max) {
+    public void setMax(double max) {
         this.max = max;
     }
